Refactor dashboard client and styles for latest layout
